Fix 3/4 speed fans
[Problem]
The Home app would allow a fan to be set to any value from 0-100%, despite a fan
being selected in the HomeBridge v2 SmartApp as a "3 speeds" or "4 speeds" fan.

[Solution]
The Characteristic.RotationSpeed was mistakenly having the "minSteps" property
set, when the actual property name is "minStep".

[Testing]
Restarted HomeBridge, and verified Home app only allows setting multiples of
33%.

issue tonesto7/homebridge-smartthings-v2#21